requirements.txt
python-telegram-bot[fast]>=21.5,<22
fastapi>=0.115
uvicorn>=0.30